# ui/valset-standardizer-frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:30066"
    }
  ],
...
  "paths": {
...
    "/api/transfer-run-logs": {
      "get": {
        "summary": "查询文件收发运行日志列表。",
        "deprecated": false,
        "tags": [
          "TransferRunLogController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MultiResultTransferRunLogViewDTO"
                }
              }
            }
          }
        },
        "operationId": "listLogs",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源主键",
            "required": false,
            "name": "sourceId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "文件主键",
            "required": false,
            "name": "transferId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "路由主键",
            "required": false,
            "name": "routeId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行阶段",
            "required": false,
            "name": "runStage"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行状态",
            "required": false,
            "name": "runStatus"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "触发类型",
            "required": false,
            "name": "triggerType"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "查询上限",
            "required": false,
            "example": 0,
            "name": "limit"
          }
        ]
      }
    },
    "/api/transfer-run-logs/page": {
      "get": {
        "summary": "分页查询文件收发运行日志。",
        "deprecated": false,
        "tags": [
          "TransferRunLogController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResultTransferRunLogViewDTO"
                }
              }
            }
          }
        },
        "operationId": "pageLogs",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源主键",
            "required": false,
            "name": "sourceId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "文件主键",
            "required": false,
            "name": "transferId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "路由主键",
            "required": false,
            "name": "routeId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行阶段",
            "required": false,
            "name": "runStage"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行状态",
            "required": false,
            "name": "runStatus"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "触发类型",
            "required": false,
            "name": "triggerType"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "关键字",
            "required": false,
            "name": "keyword"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "页码",
            "required": false,
            "example": 0,
            "name": "pageIndex"
          },
          {
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "in": "query",
            "description": "每页条数",
            "required": false,
            "example": 0,
            "name": "pageSize"
          }
        ]
      }
    },
    "/api/transfer-run-logs/analysis": {
      "get": {
        "summary": "统计分析文件收发运行日志。",
        "deprecated": false,
        "tags": [
          "TransferRunLogController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResultTransferRunLogAnalysisViewDTO"
                }
              }
            }
          }
        },
        "operationId": "analyzeLogs",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "来源主键",
            "required": false,
            "name": "sourceId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "文件主键",
            "required": false,
            "name": "transferId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "路由主键",
            "required": false,
            "name": "routeId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行阶段",
            "required": false,
            "name": "runStage"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "运行状态",
            "required": false,
            "name": "runStatus"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "触发类型",
            "required": false,
            "name": "triggerType"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "关键字",
            "required": false,
            "name": "keyword"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TransferRunLogStageAnalysisViewDTO": {
        "type": "object",
        "properties": {
          "runStage": {
            "format": "string",
            "description": "运行阶段。",
            "type": "string"
          },
          "stageLabel": {
            "format": "string",
            "description": "阶段名称。",
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "description": "阶段日志总数。",
            "type": "number",
            "example": 0
          },
          "statusCounts": {
            "description": "阶段下的状态统计。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferRunLogStatusCountViewDTO"
            }
          }
        }
      },
...
      "TransferRunLogStatusCountViewDTO": {
        "type": "object",
        "properties": {
          "runStatus": {
            "format": "string",
            "description": "运行状态。",
            "type": "string"
          },
          "statusLabel": {
            "format": "string",
            "description": "状态名称。",
            "type": "string"
          },
          "count": {
            "format": "int64",
            "description": "状态数量。",
            "type": "number",
            "example": 0
          }
        }
      },
...
      "SingleResultTransferRunLogAnalysisViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回的单条数据(object)",
            "$ref": "#/components/schemas/TransferRunLogAnalysisViewDTO"
          }
        }
      },
...
      "TransferRunLogViewDTO": {
        "type": "object",
        "properties": {
          "runLogId": {
            "format": "string",
            "description": "运行日志主键。",
            "type": "string"
          },
          "sourceId": {
            "format": "string",
            "description": "来源主键。",
            "type": "string"
          },
          "sourceType": {
            "format": "string",
            "description": "来源类型。",
            "type": "string"
          },
          "sourceCode": {
            "format": "string",
            "description": "来源编码。",
            "type": "string"
          },
          "sourceName": {
            "format": "string",
            "description": "来源名称。",
            "type": "string"
          },
          "routeName": {
            "format": "string",
            "description": "路由名称。",
            "type": "string"
          },
          "targetName": {
            "format": "string",
            "description": "目标名称。",
            "type": "string"
          },
          "transferId": {
            "format": "string",
            "description": "文件主键。",
            "type": "string"
          },
          "originalName": {
            "format": "string",
            "description": "原始文件名称。",
            "type": "string"
          },
          "routeId": {
            "format": "string",
            "description": "路由主键。",
            "type": "string"
          },
          "triggerType": {
            "format": "string",
            "description": "触发类型。",
            "type": "string"
          },
          "runStage": {
            "format": "string",
            "description": "运行阶段。",
            "type": "string"
          },
          "runStatus": {
            "format": "string",
            "description": "运行状态。",
            "type": "string"
          },
          "logMessage": {
            "format": "string",
            "description": "运行说明。",
            "type": "string"
          },
          "errorMessage": {
            "format": "string",
            "description": "错误信息。",
            "type": "string"
          },
          "createdAt": {
            "format": "string",
            "description": "创建时间。",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
      "MultiResultTransferRunLogViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferRunLogViewDTO"
            }
          }
        }
      },
...
      "TransferRunLogAnalysisViewDTO": {
        "type": "object",
        "properties": {
          "totalCount": {
            "format": "int64",
            "description": "日志总数。",
            "type": "number",
            "example": 0
          },
          "sourceCount": {
            "format": "int64",
            "description": "来源数量。",
            "type": "number",
            "example": 0
          },
          "routeCount": {
            "format": "int64",
            "description": "路由数量。",
            "type": "number",
            "example": 0
          },
          "targetCount": {
            "format": "int64",
            "description": "目标数量。",
            "type": "number",
            "example": 0
          },
          "stageAnalyses": {
            "description": "阶段统计列表。",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferRunLogStageAnalysisViewDTO"
            }
          }
        }
      },
...
      "PageResultTransferRunLogViewDTO": {
        "type": "object",
        "properties": {
          "success": {
            "format": "boolean",
            "description": "状态",
            "type": "boolean",
            "example": true
          },
          "dataType": {
            "format": "string",
            "description": "返回数据格式",
            "type": "string"
          },
          "code": {
            "description": "返回状态码(object)",
            "type": "object"
          },
          "message": {
            "format": "string",
            "description": "返回消息",
            "type": "string"
          },
          "tips": {
            "format": "string",
            "description": "返回提示消息",
            "type": "string"
          },
          "totalCount": {
            "format": "int64",
            "description": "总数",
            "type": "number",
            "example": 0
          },
          "pageSize": {
            "format": "int32",
            "description": "页数",
            "type": "integer",
            "example": 0
          },
          "pageIndex": {
            "format": "int32",
            "description": "当前页数",
            "type": "integer",
            "example": 0
          },
          "data": {
            "description": "返回数据",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferRunLogViewDTO"
            }
          }
        }
      },
...
    }
  }
}